+++ geotrek/infrastructure/views.py
@@ -29,7 +29,8 @@ class InfrastructureFormatList(MapEntityFormat, InfrastructureList):
     default_extra_columns = [
         'id', 'name', 'type', 'condition', 'description', 'accessibility',
         'implantation_year', 'published', 'publication_date', 'structure', 'date_insert',
-        'date_update', 'cities', 'districts', 'areas', 'usage_difficulty', 'maintenance_difficulty', 'uuid',
+        'date_update', 'cities', 'districts', 'areas', 'usage_difficulty',
+        'maintenance_difficulty', 'access', 'uuid',
     ] + AltimetryMixin.COLUMNS
 
 
@@ -81,7 +82,7 @@ class InfrastructureViewSet(GeotrekMapentityViewSet):
             qs = qs.annotate(api_geom=Transform('geom', settings.API_SRID))
             qs = qs.only('id', 'name', 'published')
         else:
-            qs = qs.select_related('type', 'maintenance_difficulty', 'usage_difficulty')
+            qs = qs.select_related('type', 'condition', 'maintenance_difficulty', 'access', 'usage_difficulty')
         return qs